Create Pressable and PressableState
These two classes will be used to represent objects which can be pressed and
released, kind of like a button. Later, I am going to add a Button class, which
will implement the Pressable interface. However, I do not think it will have
any other functionality but to be a hierarchical divider. I may change my mind
about this later on. But for now, I like these names.