Create InputDevice class
This commit introduces the InputDevice class, which will be the base class for
any device which sends input to the application (e.g., keyboards, mouses, etc.)
The commit also introduces that InputDevice depends on, like DeviceButton and
the ButtonPresent annotation. DeviceButton is simply an extension of the Button
class. The ButtonPresent annotation is a utility annotation, functioning in a
similar (but not identical) way to Program's Uniform annotation.