Create DeviceAdapter class
This class will be used to map input data from a source to an input device.
The idea behind this will be to allow different APIs like GLFW or XInput to
be used, while allowing the same controller class to be used unchanged. Not
only will this provide portability, but also allow features (such as rumble)
to be enabled without the need to update other parts of the program. This
commit also adds the MappedButton class and ButtonMapping annotation. The
MappedButton class will be used to map device buttons. The ButtonMapping
annotation will be used in fashion similar to the ButtonPresent annotation.