Add AnalogAdapter and ButtonAdapter annotation
These annotations make a much better way to adapt for specific types of analog
mappings and button mappings for device adapters. Rather than use a million
templates, methods can simply be marked with these annotations to state that
they can adapt for specific types. The outcome, which is seen in the updated
GLFWJoystickAdapter and GLFWXboxControllerAdapter classes, is cleaner code.