Begin transition to InputSource system
This new system of "input sources" aims to reduce redundant code that was
becoming increasingly apparent with buttons and analog values. This new system
works by having each input source start with an initial value. For example, a
button initially be unpressed; while an analog stick would be (0, 0, 0). In
future commits, device adapters will be updated to support this new system of
generic input sources.