Begin implementing InputSource based DeviceAdapter
This commits overhauls the DeviceAdapter class to work with the InputSource
concept introduced in the previous commit. Device adapters work with input
sources by having methods which take in a mapping and updating the current
value of an input source. These methods are known as "source adapters", and
are annotated with the SourceAdapter annotation (which was also added in this
commit). Later commits plan on adding finishing touches to this system, along
with updates to the JavaDocs.