Rebrand "source" to "feature" in input system
When implementing code to enable features that require output to input devices
(such as rumble for XBOX), I found that most of the code remained the same. So
as to prevent myself from writing redundant code, I renamed the  InputSource
class DeviceFeature. Most of the changes in this commit simply rename classes,
rename functions, and update JavaDocs to reflect this commit. Future commits
will have the rumble feature implemented for the XBOX controller.