Create DeviceSeeker and X-input device seeker
The purpouse of a DeviceSeeker is to scan the system for input devices that
should be recognized by the input system. Likewise, it should ensure that
connected devices are still connected and disconnect them if they are not.
The implementation of this system is one of the final parts of the grand
plan of the input API, and I am greatly pleased with how it is coming out.