Add uniform shortcut methods to Controller
These shortcut methods (e.g., getLeftAnalog()) provide a convenient way to
access data that is usually present on a controller. By deisgn, his will not
be extended to buttons by design (e.g., there will be no isAPressed()). For
buttons, uniformity will be accomplished via an "action" system. This has yet
to be implemented, but will be added in later commits.