Delete DeviceListener
I didn't like how this system was coming out, so I opted to switch to the same
event system that the Audio system uses. The next commits should introduce the
code needed for this. I'll also probably update Whirvex Event later to have a
bit more flexibilty and stability than it currently has.