Start implementing Keyboard support
This commit adds a big chunk of the code needed to add support for the Keyboard
and Mouse input devices. While this commit doesn't fully implement everything,
I decided to make this commit now so this chunk of code didn't get tangled in
other big changes I need to make to the input system.