Introduce feature monitor system
This system comes in handy for preventing duplicate code from being written
across input devices which are different but have similar (or identical) forms
of input. For example, a keyboard vs. a controller. While the only current
implementation of a FeatureMonitor is the ConnectionMonitor, which may appear
as an over-engineered implementation, I can assure that future commits will
contain other FeatureMonitor implementation which will show its usefulness.