Add LS/RS and LT/RT trigger fields for Controller
I decided to add these fields as it would make it a whole lot easier to grab
the specific LS/RS and LT/RT device features of a controller, rather than just
their values. It also gets rid of the redundant code for getting said values
found in the classes which extending from Controller.