Add support for DualShock 4 controllers via HID
I am really excited for this, as it enables possibilities that were previously
not possible for PS4 controllers using GLFW! Using the HID adapters, it is now
possible to make the controllers rumble. It is also possible to change the
lightbar color! I couldn't get the speaker to work for this commit (believe me,
I tried.) However, I intend on getting it working at some point in the future.