Make HidDeviceSeeker set HID devices to non-blocking
I made this change as I found myself writing the same code for the Wiimote HID
adapter (which will be added in later commits.) So, I just went on and made the
HID device seeker set the devices to non-blocking. The Ds4HidAdapter will also
no longer open the HID device when constructed. That is the responsibility of
the one who owns the HID device.