Use LibUsb instead of JavaX for USB communications
I decided to make this change as it gives me more control when writing code
for USB devices. This is necessary for me to add support for PS3 controllers.
They require communications to take place which the JavaX USB API simply did
not support. I did not want to have a mix of high level and low level USB
communication in my project. As such, I switched to LibUsb for the Nintendo
GameCube controllers first in this commit. Support for the PlayStation 3
controller will be added in the following commits.