Add final poll() call for unregistered devices in DeviceSeeker
I discovered that devices which were being unregistered failed to report that
they had disconnected via the events system. This was because when they were
unregistered (due to disconnection), their ConnectionMonitor feature never got
to take a poll. As such, it never got to see its assigned device was no longer
connected and fire the DeviceDisconnectEvent. This commit fixes that by adding
a final poll to the device when it is unregistered.