Quit using Whirvex Event
I decided to remove the usage of Whirvex Event from this API. This is for a
multitude of reasons. The main one being I have grown to dislike Whirvex Event
as a whole, and believe that callbacks would be more suitable for Kibasan. I
will implement these callbacks in later commits.