Add Mockito Core to dependences in X-input module
I'm hoping this will fix the broken builds in Jenkins. While building
on local machines already works, for some reason it does not in Jenkins.
Build fails as soon as it reaches testing for the X-input module, with
Mockito stating it can't mock a class for some reason. Sources have
suggested that I add this dependency to possibly fix the build.