Create LibUsbDevice to wrap LibUSB calls
It was proving difficult to write unit testing for LibUsbDeviceSeeker, as it
made calls to native methods. Mockito does not support mocking these. To get
around this, I made a wrapper class. This has actually had the side effect
of making the overall code in LibUsbDeviceSeeker look cleaner! Later commits
will be updates to the LibUsbDeviceSeeker which make use of this class.