Use new IoFeature module
This new model utilizes what are called "internal states" and "container
states". The internal state is visible only to the I/O device and adapter.
The container state is visible publicly. This model enables more complex
state models for I/O features. It also allows for the read-only interface
hacky workaround to be done away with. This commit resulted in a lot of
tests breaking, mainly in the devices/ modules. I will create new unit
tests that match this new system in later commits.