Finish transition to RxJava in API module, add update() to IoFeature
This commit transations events over to the RxJava API, which speciailizes in
reactive programming. Although it is only used as an event bus, it is a much
needed improvement over the limited callbacks originally present in this
project. The update() method in IoFeature will also replace the clunky and
error prone feature monitors found in other modules. Later commits will fix
compile errors caused by the changes in this commit.