Require feature IDs to be unique, add accompanying utility methods
This commit requires that all registered features in a feature registry to
have unique IDs. This allows for methods like isFeatureWithIdRegistered()
and getFeatureById() to exist. These should make it easy to store data for
I/O features, using their ID as the key, to storage. An example would be
analog stick calibrations in a configuration file.