Remove hid-usb modules from master branch
This is only a temporary measure. Since unit testing has proven difficult
for these modules, I have opted to create a dedicated branch for them so
I can work on them separately. If they can't be finished in time for the
first release, they will simply be added in a later version.