Add GenericSensor feature to controller module
This is just a placeholder for now. I wanted this features to be declared
in their respective devices, even if they have no adapters which support
them in v1.0.0 of Ketill. If any changes need to be made to this feature
or its state class, I will do so in a later commit.