Introduce new I/O feature system.
This new system has new classes such as IoState, IoFlow, and IoLogic.
These work to address the issues of the old system, by re-structuring
the state system, autonomy, as well as introducing the brand new idea
of an IoFeature having a flow. That being, it only goes in a certain
direction as opposed to all of their states being updated at poll.